Refactor API with DTOs and CRUD operations
Introduce request/response DTOs to separate API concerns from domain models. Add comprehensive CRUD endpoints for courses and students through CourseController and StudentController. Create CourseMapper for entity-DTO conversions and enhance StudentMapper. Implement AdminService and AdminController for database management. Add new repository query methods for course-student relationships and update entity mappings to support nullable course assignments.

+++ SpringBoot/JpaMapping/src/main/java/learning/springboot/JpaMapping/model/Course.java
@@ -15,7 +15,7 @@ import java.util.List;
 public class Course {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long course_id;
+    private int course_id;
     private String title;
     private String code;
     private String staff;
